feat: 1. navigation 2. path alias
