+++ apps/muse-mobile/android/app/src/main/java/com/musemobile/MainActivity.java
@@ -1,4 +1,5 @@
 package com.musemobile;
+import android.os.Bundle;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -16,6 +17,10 @@ public class MainActivity extends ReactActivity {
     return "MuseMobile";
   }
 
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+    super.onCreate(null);
+  }
   /**
    * Returns the instance of the {@link ReactActivityDelegate}. Here we use a util class {@link
    * DefaultReactActivityDelegate} which allows you to easily enable Fabric and Concurrent React
