+++ apps/muse-mobile/package.json
@@ -4,15 +4,19 @@
   "private": true,
   "dependencies": {
     "@react-native-async-storage/async-storage": "*",
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
     "@rnx-kit/metro-resolver-symlinks": "^0.1.26",
+    "@sentry/react-native": "5.1.1",
     "@testing-library/jest-native": "*",
     "@testing-library/react-native": "*",
+    "babel-plugin-module-resolver": "*",
+    "metro-config": "*",
     "react": "*",
     "react-native": "*",
     "react-native-config": "*",
-    "react-native-svg": "*",
-    "@sentry/react-native": "5.1.1",
-    "metro-config": "*",
-    "babel-plugin-module-resolver": "*"
+    "react-native-safe-area-context": "^4.5.0",
+    "react-native-screens": "^3.20.0",
+    "react-native-svg": "*"
   }
 }
