--- apps/muse-mobile/src/app/App.spec.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react-native';
-
-import App from './App';
-
-test('renders correctly', () => {
-  const { getByTestId } = render(<App />);
-  expect(getByTestId('heading')).toHaveTextContent('Welcome');
-});
