+++ libs/types/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/common';
+export * from './mobile';
