+++ libs/types/src/mobile/index.ts
@@ -1,0 +1,1 @@
+export * from './routeName';
