+++ libs/types/src/mobile/routeName.ts
@@ -1,0 +1,3 @@
+export enum RouteNames {
+  Home = 'Home',
+}
